Java 5 type checking and compiler warnings
git-svn-id: svn://source.pentaho.org/svnkettleroot/Kettle/trunk@3722 5fb7f6ec-07c1-534a-b4ca-9155e429e800

+++ src/org/pentaho/di/core/util/Base64.java
@@ -724,7 +724,7 @@ public class Base64
         if( bytes != null && bytes.length >= 4 )
         {
             
-            int head = ((int)bytes[0] & 0xff) | ((bytes[1] << 8) & 0xff00);       
+            int head = (bytes[0] & 0xff) | ((bytes[1] << 8) & 0xff00);       
             if( java.util.zip.GZIPInputStream.GZIP_MAGIC == head ) 
             {
                 java.io.ByteArrayInputStream  bais = null;
